Implement stub testing for functions in dao.js
In dao.js, the query functions have an optional second boolean parameter called stub, which is automatically set to false. In test.js, we can now specify that we want to call a function in stub mode by passing true as the second argument to any of the query functions. For example insert(data, true) will return the contents of the data variable. This allows us to test that our functions are passed the proper parameters.